Fixes to allow input files larger than single threaded size

+++ compression.h
@@ -5,6 +5,6 @@
 
 size_t compression_key_size(size_t length);
 size_t compression_compress(Key key, uint8_t length, char* buffer, uint8_t* places);
-Key compression_decompress(char* buffer, size_t length);
+Key compression_decompress(char* buffer, uint8_t length);
 
 #endif
